Fold countdown module into voting round as private implementation
Move arm, cancel, and evaluate from convex/model/countdown.ts into
convex/model/votingRound.ts as private (non-exported) functions. The
countdown had exactly one adapter (the voting round module), making it
a hypothetical seam — not a real abstraction boundary.

Changes:
- Folding arm/cancel/evaluate into votingRound.ts removes the 70-line
  countdown.ts file and the cross-file import
- All 5 Countdown.cancel and 3 Countdown.evaluate call sites now call
  the private helpers directly
- cancelCountdown exported function continues to delegate to the
  private cancel helper, preserving the public interface
- Tests: removed standalone Countdown.arm/cancel/evaluate describe
  blocks; replaced direct Countdown calls with armCountdown helper
  that exercises the round's public castVote interface
- Removed unused deleteVote test helper

The round module deepens: the entire auto-reveal lifecycle (arm,
cancel, evaluate, scheduled reveal, token check) now lives in one
file. All 160 tests pass.

Closes #219
